add fishing mechanic & luck

+++ Assets/Scripts/Fishing/Player & Autofisher Stats/SO_Stats.cs
@@ -6,7 +6,7 @@ public class SO_Stats : ScriptableObject
     public string nameID; 
     public int luck;
     public float fishSpeed;
-    public float skill;
+    public int skill;
 
     public GameObject Prefab;
 }
